Add default file matching for other common file names with this syntax
As commit fd7de913 removed general *.txt file matching,
some file names such as "more_requirements.txt" and constraints files "constraints.txt"
will not be detected by default by this plugin. This commit remedies that,
adding patterns that will match all example file names from
https://pip.pypa.io/en/stable/reference/pip_install/#requirements-file-format

Constraints files are another feature of pip. The difference from requirements files
is that requirements added in a constraints file won't cause the requirement to be installed.
The constraints file just specifies which versions of that package are allowed to be installed if
the package was to be installed. Constraints files are used by adding the "-c <filename>" option to a pip install command.
As constraints files use the same syntax as requirements files, these can be handled by the same plugin.

The reason for keeping the exact string matching "requirements.txt" in addition to the newly-added
wildcard match "*requirements.txt" is to allow the user to easily remove the wildcard but keep the
exact match in the case that this behaviour is inconvenient. The plugin would match the same files without
the exact string match.